requirements.txt
pywin32
psutil
PySide6
google-generativeai
Pillow
pyinstaller
pynput
python-dotenv
